+++ protocols/CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GetIdentity endpoint to auth.
 - `Login` that performs an interactive login. This is done with a stream of
   `InteractiveLoginCommand` that the client follows.
+- `WaitForRemoteAccessRequest` endpoint that can be used to wait for approval of a remote
+  access request. This should be used together with gRPC timeouts.
+- `CancelRemoteAccessReqest` endpoint to remove a pending remote access request.
 
 ## [1.0.0] - 2021-07-03